Move unit terminology from assets to modules/_shared folder
